+++ src/app.py
@@ -30,157 +30,10 @@ from flask import request
 from dash.exceptions import PreventUpdate
 from dash.dependencies import Input, Output, State, ALL, MATCH
 
-# ----------------------------------------------------------------------------
-# CONFIG SETTINGS
-# ----------------------------------------------------------------------------
-DATA_PATH = pathlib.Path(__file__).parent.joinpath("data")
-ASSETS_PATH = pathlib.Path(__file__).parent.joinpath("assets")
-REQUESTS_PATHNAME_PREFIX = os.environ.get("REQUESTS_PATHNAME_PREFIX", "/")
+# import local modules
+from config_settings import *
+from data_processing import *
 
-# ----------------------------------------------------------------------------
-# SECURITY FUNCTION
-# ----------------------------------------------------------------------------
-def get_django_user():
-    """
-    Utility function to retrieve logged in username
-    from Django
-    """
-    DJANGO_LOGIN_API = os.environ.get("DJANGO_LOGIN_API", None)
-    DJANGO_SESSION_COOKIE = os.environ.get("DJANGO_SESSION_COOKIE", None)
-    try:
-        if not DJANGO_LOGIN_API:
-            return True
-        if not DJANGO_SESSION_COOKIE:
-            raise Exception("DJANGO_SESSION_COOKIE environment variable missing")
-        if not request:
-            raise Exception("Request context missing while trying to authorize user")
-        session_id = request.cookies.get(DJANGO_SESSION_COOKIE)
-        if not session_id:
-            print("Request cookies: ", request.cookies)
-            raise Exception("{cookie} cookie is missing".format(cookie=DJANGO_SESSION_COOKIE))
-        api = "{django_login_api}".format(
-            django_login_api=DJANGO_LOGIN_API
-        )
-        response = requests.get(
-            api, 
-            headers = {
-                'cookie': '{cookie}={session_id}'.format(
-                    cookie=DJANGO_SESSION_COOKIE,
-                    session_id=session_id
-                )
-            }
-        )
-        return response.json()
-    except Exception as e:
-        print(e)
-        return None
-
-# ----------------------------------------------------------------------------
-# STYLING
-# ----------------------------------------------------------------------------
-
-TACC_IFRAME_SIZE = {
-    "max-width" : "1060px", "max-height" : "980px" # THESE ARE SET TO FIT IN THE 1080x1000 TACC iFRAME.  CAN BE REMOVED IF THOSE CONSTRAINTS COME OFF
-}
-
-CONTENT_STYLE = {
-    "padding": "2rem 1rem",
-    "font-family": 'Arial, Helvetica, sans-serif',
-}
-
-export_style = '''
-    position:absolute;
-    right:25px;
-    bottom:-55px;
-    font-family: Arial, Helvetica, sans-serif;
-    margin: 10px;
-    color: #fff;
-    background-color: #17a2b8;
-    border-color: #17a2b8;
-    display: inline-block;
-    font-weight: 400;
-    text-align: center;
-    white-space: nowrap;
-    vertical-align: middle;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-    border: 1px solid transparent;
-    padding: .375rem .75rem;
-    font-size: 1rem;
-    line-height: 1.5;
-    border-radius: .25rem;
-    transition: color .15s ease-in-out,background-color .15s ease-in-out,border-color .15s ease-in-out,box-shadow .15s ease-in-out;
-'''
-
-
-# ----------------------------------------------------------------------------
-# API DATA FUNCTIONS
-# ----------------------------------------------------------------------------
-
-## FUNCTIONS FOR LOADING DATA: MOVE TO MODULE
-def load_api_data(api_url):
-    '''Get dictionary of sankey data if provided a valid API url'''
-    try:
-        response = requests.get(api_url)
-        api_data = response.json()
-    except:
-        return False
-    return api_data
-
-def get_api_df(api_data):
-    ''' Take a dictionary of data balues and return a dataframe with source, target and value columns to build sankey '''
-    if load_api_data(api_url):
-        api_data = load_api_data(api_url)
-        df = pd.DataFrame.from_dict(api_data)
-    else:
-        cosort_columns = ['source','target','value']
-        df = pd.DataFrame(columns = cosort_columns)
-    return df
-
-def get_sankey_nodes(dataframe,source_col = 'source', target_col = 'target'):
-    ''' Extract node infomration from sankey dataframe in case this is not provided '''
-    nodes = pd.DataFrame(list(dataframe[source_col].unique()) + list(dataframe[target_col].unique())).drop_duplicates().reset_index(drop=True)
-    nodes.reset_index(inplace=True)
-    nodes.columns = ['NodeID','Node']
-    return nodes
-
-def get_sankey_dataframe (data_dataframe,
-                          node_id_col = 'NodeID', node_name_col = 'Node',
-                          source_col = 'source', target_col = 'target', value_col = 'value'):
-    ''' Merge Node dataframes with data dataframe to create dataframe properly formatted for Sankey diagram.
-        This means each source and target gets assigned the Index value from the nodes dataframe for the diagram.
-    '''
-    # get nodes from data
-    nodes = get_sankey_nodes(data_dataframe)
-
-    # Copy of Node data to merge on source
-    sources = nodes.copy()
-    sources.columns = ['sourceID','source']
-
-    # Copy of Node data to merge on target
-    targets = nodes.copy()
-    targets.columns = ['targetID','target']
-
-    # Merge the data dataframe with node information
-    sankey_dataframe = data_dataframe.merge(sources, on='source')
-    sankey_dataframe = sankey_dataframe.merge(targets, on='target')
-
-    return nodes, sankey_dataframe
-
-def load_historical_data(csv_url):
-    '''Load csv of historical dates and extract unique dates for dropdown options'''
-    try :
-        # Data Frame
-        historical_data = pd.read_csv(csv_url, header = None)
-        historical_data.columns=['source','target','value','date']
-        historical_data['date_time'] = historical_data['date'].apply(lambda x: pd.Timestamp(x).strftime('%Y-%m-%d (%H:%M)'))
-
-    except :
-        historical_data = 'Could not access data'
-
-    return historical_data
 
 # ----------------------------------------------------------------------------
 # DATA urls
